Audit Org corrections

+++ foodbank-rest-api/src/main/java/foodbank/it/persistence/repository/IAuditAssocRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import foodbank.it.persistence.model.AuditAssoc;
 
-public interface IAuditAssocRepository  extends PagingAndSortingRepository<AuditAssoc, Integer>{
+public interface IAuditAssocRepository  extends PagingAndSortingRepository<AuditAssoc, Long>{
 	
 }
